// components/MovesUs/types/moveUsTypes.ts
import { ImageData } from "@/helpers/generalTypes";

export interface MoveUsTypes {
  tag: string;
  englishTag: string;
  title: string;
  englishTitle: string;
  ourHistoryTitle: string;
  ourHistoryTitleEnglish: string;
  ourHistoryText: string;
  ourHistoryTextEnglish: string;
  image: ImageData;
}