// components/Newsletter/types/newsletterTypes.ts
import { ButtonLink } from "@/helpers/generalTypes";

export interface NewsLetterTypes {
  tag: string;
  title: string;
  englishTitle: string;
  subtitle: string;
  englishSubtitle: string;
  placeholderInput: string;
  englishPlaceholderInput: string;
  textButton: string;
  englishTextButton: string;
}